Check if temporary directory exists before trying to remove it in Layout.import_action()

+++ code/Layout/Layout.js
@@ -268,7 +268,7 @@ Layout.prototype.import_action = function() {
       } catch (ex) {
          res.message = ex;
          app.log(ex);
-         temp.removeDirectory();
+         temp && temp.removeDirectory();
          res.redirect(this.href(req.action));
       }
       res.redirect(this.href());
